refactor: rename PageSystems to SolarSystems; move all canvas related files to sensical locations
