--- src/presentation/components/canvas/Canvas.styles.tsx
@@ -1,7 +1,0 @@
-import { css } from '@emotion/react';
-
-export const styles = css`
-  container: {
-    width: '100%',
-  }
-`;
