+++ src/presentation/components/canvas/Canvas/BackgroundPoint/IStarPoint.ts
