+++ src/presentation/components/canvas/Canvas/Canvas.scss
