+++ src/presentation/components/canvas/Canvas/Canvas.styles.tsx
@@ -1,0 +1,9 @@
+import { SxProps, Theme } from '@mui/material';
+
+export const boxStyles: SxProps<Theme> = {
+  width: '100%',
+  height: '100%',
+  position: 'absolute',
+  overflow: 'hidden',
+  zIndex: -1,
+};
