+++ src/presentation/components/canvas/Canvas/StarSprite/StarSprite.ts
