+++ src/presentation/components/canvas/Canvas/renderPlanets.ts
