+++ src/presentation/components/canvas/Canvas/renderStars.ts
