+++ src/presentation/components/canvas/Canvas/renderUI.ts
