+++ src/presentation/components/canvas/Canvas/wipeCanvas.ts
