+++ src/presentation/components/canvas/index.ts
@@ -1,9 +1,9 @@
-export type { IStarField } from './IStarField';
-export type { IStarPoint } from './IStarPoint';
+export type { IStarField } from './Canvas/Background/IStarField';
+export type { IStarPoint } from './Canvas/BackgroundPoint/IStarPoint';
 
-export { Canvas } from './Canvas.component';
+export { Canvas } from './Canvas/Canvas.component';
 
-export { renderStarField } from './renderStarField';
-export { renderStars } from './renderStars';
-export { renderPlanets } from './renderPlanets';
-export { renderUI } from './renderUI';
+export { renderStarField } from './Canvas/renderStarField';
+export { renderStars } from './Canvas/renderStars';
+export { renderPlanets } from './Canvas/renderPlanets';
+export { renderUI } from './Canvas/renderUI';
