+++ src/presentation/components/pages/SolarSystems/SolarSystems.scss
