+++ src/presentation/components/pages/SolarSystems/SolarSystems.spec.tsx
@@ -1,0 +1,6 @@
+import { cleanup, render } from '@testing-library/react';
+import { SolarSystems } from './SolarSystems.component';
+
+describe('<SolarSystems />', () => {
+  afterEach(cleanup);
+});
