+++ src/presentation/components/pages/SolarSystems/SolarSystems.styles.ts
@@ -1,0 +1,6 @@
+import { SxProps, Theme } from '@mui/material';
+
+export const dividerStyles: SxProps<Theme> = {
+  marginY: 1.5,
+  backgroundColor: '#f0f0f0',
+};
