+++ src/presentation/components/pages/index.ts
@@ -3,7 +3,7 @@ export { PageFooter } from './PageFooter/PageFooter.component';
 
 import { PageHome } from './PageHome/PageHome.component';
 import { Data } from './Data/Data.component';
-import { PageSystems } from './PageSystems/PageSystems.component';
+import { SolarSystems } from './SolarSystems/SolarSystems.component';
 import { PageBookmarks } from './PageBookmarks/PageBookmarks.component';
 import { About } from './About/About.component';
 import { NotFound } from './NotFound/NotFound.component';
@@ -11,7 +11,7 @@ import { NotFound } from './NotFound/NotFound.component';
 export const Pages = {
   PageHome,
   Data,
-  PageSystems,
+  SolarSystems,
   PageBookmarks,
   About,
   NotFound,
